+++ scalr/provider.go
@@ -63,14 +63,15 @@ func Provider() terraform.ResourceProvider {
 		},
 
 		DataSourcesMap: map[string]*schema.Resource{
-			"scalr_workspace":     dataSourceScalrWorkspace(),
-			"scalr_workspace_ids": dataSourceScalrWorkspaceIDs(),
-			"scalr_current_run":   dataSourceScalrCurrentRun(),
-			"scalr_endpoint":      dataSourceScalrEndpoint(),
-			"scalr_webhook":       dataSourceScalrWebhook(),
-			"scalr_environment":   dataSourceScalrEnvironment(),
-			"scalr_role":          dataSourceScalrRole(),
-			"scalr_access_policy": dataSourceScalrAccessPolicy(),
+			"scalr_workspace":      dataSourceScalrWorkspace(),
+			"scalr_workspace_ids":  dataSourceScalrWorkspaceIDs(),
+			"scalr_current_run":    dataSourceScalrCurrentRun(),
+			"scalr_endpoint":       dataSourceScalrEndpoint(),
+			"scalr_webhook":        dataSourceScalrWebhook(),
+			"scalr_environment":    dataSourceScalrEnvironment(),
+			"scalr_role":           dataSourceScalrRole(),
+			"scalr_access_policy":  dataSourceScalrAccessPolicy(),
+			"scalr_module_version": dataSourceModuleVersion(),
 		},
 
 		ResourcesMap: map[string]*schema.Resource{
@@ -79,6 +80,7 @@ func Provider() terraform.ResourceProvider {
 			"scalr_endpoint":      resourceScalrEndpoint(),
 			"scalr_webhook":       resourceScalrWebhook(),
 			"scalr_environment":   resourceScalrEnvironment(),
+			"scalr_module":        resourceScalrModule(),
 			"scalr_role":          resourceScalrRole(),
 			"scalr_access_policy": resourceScalrAccessPolicy(),
 		},